Okino813

+++ app/Http/Controllers/GroupController.php
@@ -3,6 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Group;
+use App\Models\GroupSport;
+use App\Models\Sport;
+use App\Models\User;
+use App\Models\Member;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -30,11 +34,12 @@ class GroupController extends Controller
     public function create()
     {
         $user = Auth::user();
-        if($user == null){
+        if ($user == null) {
             return view('auth.login');
         }
 
         return view('groups.create');
+
     }
 
     /**
@@ -71,7 +76,38 @@ class GroupController extends Controller
             return view('auth.login');
         }
 
-        return view('groups.show', compact('group'));
+        if($group->id != 1) {
+            return view('groups.show', compact('group'));
+        }
+        else{
+            return back();
+        }
+    }
+
+    public function showview(Group $group)
+    {
+        $user = Auth::user();
+        if($user == null){
+            return view('auth.login');
+        }
+
+        if($group->id != 1)
+        {
+            $members = Member::with(['user', 'group'])
+                    ->where('group_id', $group->id)  // Filtrer par l'ID du groupe
+                    ->get();
+
+
+            $coach = User::find($group->user_id);
+
+            $sportsGroup = $group->sports;
+
+
+            return view('groups.showview', compact('group', 'members', 'coach', 'sportsGroup'));
+        }
+        else{
+            return back();
+        }
     }
 
     /**
@@ -80,16 +116,40 @@ class GroupController extends Controller
     public function edit(Group $group)
     {
         $user = Auth::user();
-        if($user == null){
+        if ($user == null) {
             return view('auth.login');
         }
 
-        if($group->user_id == $user->id)
-        {
-            return view('dashboard.index');
+        if ($user->id = $group->user_id and $group->id != 1) {
+
+
+            // On récupère les membres
+            $members = Member::with(['user', 'group'])
+                ->where('group_id', $group->id)  // Filtrer par l'ID du groupe
+                ->get();
+
+            // Récupérer les utilisateurs qui ne sont pas membres de ce groupe
+            $membersIds = Member::where('group_id', $group->id)
+                ->pluck('user_id');  // On récupère tous les user_id des membres du groupe
+
+            // Récupérer tous les utilisateurs sauf ceux qui sont déjà membres
+            $usersNotInGroup = User::whereNotIn('id', $membersIds)->get();
+
+            $coach = User::find($group->user_id);
+
+            // On récupère tous les sports associés au groupe via la relation 'sports' (correspond à 'group_sport')
+            $sportsGroup = $group->sports;  // Utilisation de la relation 'sports' définie dans le modèle 'Group'
+
+            // Récupérer les sports qui ne sont pas encore liés à ce groupe
+            $sportIds = $sportsGroup->pluck('id');  // On récupère tous les sport_id déjà associés au groupe
+
+            // Récupérer tous les sports sauf ceux qui sont déjà liés au groupe
+            $sportsNotInGroup = Sport::whereNotIn('id', $sportIds)->get();
+
+            return view('groups.edit', compact('group', 'members', 'usersNotInGroup', 'coach', 'sportsGroup', 'sportsNotInGroup'));
         }
         else{
-            return view('groups.edit', compact('group'));
+            return view('auth.login');
         }
     }
 
@@ -98,12 +158,19 @@ class GroupController extends Controller
      */
     public function update(Request $request, Group $group)
     {
-        $validated = $request->validate([
-            'name' => 'required|string|max:255',
-        ]);
+        $user = Auth::user();
+        if ($user == null) {
+            return view('auth.login');
+        }
 
-        $group->update($validated);
-        return redirect()->route('groups.index')->with('success', 'Group mis à jour avec succès.');
+        if ($user->id = $group->user_id) {
+            $validated = $request->validate([
+                'name' => 'required|string|max:255',
+            ]);
+
+            $group->update($validated);
+            return redirect()->route('groups.index')->with('success', 'Group mis à jour avec succès.');
+        }
     }
 
     /**
@@ -111,7 +178,17 @@ class GroupController extends Controller
      */
     public function destroy(Group $group)
     {
-        $group->delete();
-        return redirect()->route('dashboard.groups.index')->with('success', 'Group mis à jour avec succès.');
+        $user = Auth::user();
+        if ($user == null) {
+            return view('auth.login');
+        }
+
+        if ($user->id = $group->user_id) {
+            $group->delete();
+            return redirect()->route('dashboard.groups.index')->with('success', 'Group mis à jour avec succès.');
+        }
+        else{
+            return back();
+        }
     }
 }
